fix(daemon): support long Unix socket paths

+++ internal/daemon/command_server.go
@@ -72,6 +72,7 @@ type CommandServer struct {
 	targetBound bool
 	listener    net.Listener
 	isAbstract  bool
+	actualPath  string
 }
 
 // NewCommandServer wires a CommandServer to the given run directory and
@@ -102,6 +103,9 @@ func (s *CommandServer) Start() error {
 		return fmt.Errorf("chmod run dir: %w", err)
 	}
 	sockPath := CommandSocketPath(s.dir)
+	if needsShortSocketPath(sockPath) {
+		return s.startWithShortSocketPath(sockPath)
+	}
 	_ = os.Remove(sockPath)
 	listener, err := net.Listen("unix", sockPath)
 	if err != nil {
@@ -125,6 +129,30 @@ func (s *CommandServer) Start() error {
 		return fmt.Errorf("chmod command socket: %w", err)
 	}
 	s.listener = listener
+	s.actualPath = sockPath
+	go s.acceptLoop()
+	return nil
+}
+
+func (s *CommandServer) startWithShortSocketPath(logicalPath string) error {
+	actualPath := shortSocketPath(logicalPath)
+	_ = removeSocketPath(actualPath)
+	listener, err := net.Listen("unix", actualPath)
+	if err != nil {
+		return fmt.Errorf("create short command socket: %w", err)
+	}
+	if err := os.Chmod(actualPath, 0o600); err != nil {
+		_ = listener.Close()
+		_ = removeSocketPath(actualPath)
+		return fmt.Errorf("chmod command socket: %w", err)
+	}
+	if err := linkShortSocket(logicalPath, actualPath); err != nil {
+		_ = listener.Close()
+		_ = removeSocketPath(actualPath)
+		return fmt.Errorf("link command socket: %w", err)
+	}
+	s.listener = listener
+	s.actualPath = actualPath
 	go s.acceptLoop()
 	return nil
 }
@@ -153,8 +181,13 @@ func (s *CommandServer) Stop() error {
 		err = s.listener.Close()
 	}
 	if !s.isAbstract {
-		if rmErr := os.Remove(CommandSocketPath(s.dir)); rmErr != nil && !os.IsNotExist(rmErr) && err == nil {
+		if rmErr := removeSocketPath(CommandSocketPath(s.dir)); rmErr != nil && err == nil {
 			err = rmErr
+		}
+		if s.actualPath != "" && s.actualPath != CommandSocketPath(s.dir) {
+			if rmErr := removeSocketPath(s.actualPath); rmErr != nil && err == nil {
+				err = rmErr
+			}
 		}
 	}
 	return err
